// src/frontend/src/modules/identity/api/generated/api-client.ts
export interface AuthResponseDto {
    success?: boolean;
    token?: string | undefined;
    message?: string | undefined;
    user?: UserDto;
}

export interface LoginDto {
    emailOrUsername?: string | undefined;
    password?: string | undefined;
    rememberMe?: boolean;
}

export interface RegisterDto {
    username?: string | undefined;
    email?: string | undefined;
    phoneNumber?: string | undefined;
    password?: string | undefined;
}

export interface UserDto {
    id?: string | undefined;
    username?: string | undefined;
    email?: string | undefined;
}